# docs/swagger.yaml
basePath: /api/v1
definitions:
  GoFrioCalor_internal_dto.DeliveryInfoDTO:
    properties:
      cantidad:
        type: integer
      fecha_accion:
        type: string
      id:
        type: integer
      nro_cta:
        type: string
      nro_rto:
        type: string
      tipo_entrega:
        type: string
    type: object
...
  GoFrioCalor_internal_dto.DispenserInfoDTO:
    properties:
      id:
        type: integer
      marca:
        type: string
      nro_serie:
        type: string
      tipo:
        type: string
      validated:
        type: boolean
    type: object
...
  GoFrioCalor_internal_dto.ValidateTokenRequest:
    properties:
      fecha_accion:
        type: string
      nro_cta:
        type: string
      token:
        minLength: 4
        type: string
    required:
    - fecha_accion
    - nro_cta
    - token
    type: object
  GoFrioCalor_internal_dto.ValidateTokenResponse:
    properties:
      delivery:
        $ref: '#/definitions/GoFrioCalor_internal_dto.DeliveryInfoDTO'
      dispensers:
        items:
          $ref: '#/definitions/GoFrioCalor_internal_dto.DispenserInfoDTO'
        type: array
      message:
        type: string
      valid:
        type: boolean
    type: object
...
host: localhost:8095
info:
  contact:
    email: soporte@gofrocalor.com
    name: Soporte GoFrioCalor
  description: |-
    API para gestión de entregas, dispensers y órdenes de trabajo de GoFrioCalor
    Esta API permite gestionar el ciclo completo de entregas, desde la creación hasta la finalización,
    incluyendo validación de términos y condiciones, integración con Infobip y flujos móviles.
  license:
    name: Propietario
    url: https://gofrocalor.com/license
  title: GoFrioCalor API
  version: "1.0"
paths:
...
  /mobile/validate-token:
    post:
      consumes:
      - application/json
      description: El repartidor valida el token proporcionado por el cliente
      parameters:
      - description: Token del cliente
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/GoFrioCalor_internal_dto.ValidateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoFrioCalor_internal_dto.ValidateTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Validar token de entrega
      tags:
      - Mobile
...
schemes:
- http
- https
swagger: "2.0"
...